Now WITH only accepts with expanders.
The reason:

Being WITH a mix of LET and MULTIPLE-VALUE-BIND felt wrong.
Each expression must do a specific thing.
Also, you couldn't know if a form will be evaluated o expanded
and I think that can be a source of unwanted errors.
Now is clearer what WITH is designed for. It is more compact.